Update Model Values - Valid Alexa build
This file will build successfully upon upload to Alexa platform.  Changed strategy on how entity resolution value is resolved.
The original strategy was to place the "spoken value" in the "value" field but the field will now contain the "written value" i.e. S90 versus S 90 and a synonym is created for the spoken version: S 90 -with a space between. The issue resolves the ability to send back the original written value the API is expecting.  As the Id field in Alexa cannot contain spaces.